backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg[binary]==3.2.10
pydantic==2.9.2
bcrypt==4.2.0
python-jose==3.3.0
requests==2.32.3
redis==5.0.8
openai==1.51.0
email-validator==2.2.0
python-dotenv==1.0.1
matplotlib==3.9.2
numpy==2.1.2
sympy==1.13.1
paypalrestsdk==1.13.1
httpx==0.27.2
alembic==1.13.1
slowapi==0.1.9
python-multipart==0.0.9
